requirements.txt
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.63
botocore==1.34.63
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jmespath==1.0.1
openai==1.14.0
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
